btcmines: libmcmap: apparently you can't just say 'pointer = pointer+0x05' to move it 5 bytes forward ... chunk data now actually points to the correct location immediately after the 5-byte chunk header

+++ libmcmap/test.c
@@ -26,7 +26,7 @@ int main(int argc, char **argv)
 				{
 				fprintf(stdout,"stats for chunk (%u,%u):\n",x,z);
 				fprintf(stdout,"\ttimestamp: %s",ctime(&(r->dates[z][x])));
-				fprintf(stdout,"\tbyte size: %ld\n",r->chunks[z][x].size);
+				fprintf(stdout,"\tbyte size: %d\n",(int)r->chunks[z][x].size);
 				}
 			else
 				fprintf(stdout,"chunk (%u,%u) does not exist\n",x,z);
